implemented switch statements
replaced the IR commands with switch statements to make the program faster, and replaced variables with defines so that the variable memory is better used

+++ L293D_constructor/Differential_Drive.h
@@ -11,7 +11,7 @@ class Differential_Drive {
     this->m1 = m1;
     this->m2 = m2;
   }
-  void drive(double* controls) {
+  void drive(int* controls) {
     m1.setSpeed(controls[0]-controls[1]);
     m2.setSpeed(controls[0]+controls[1]);
   }
